Fixed a compile error with gcc 7.2
gcc 7.2 emits a warning when memset is used without a multiplication with the size of an object and due to `-Werror` it is not possible to compile this way. In `matrix.c` `memset` is used like this. I fixed the problem by multiplying the fieldsize with `sizeof(matrix_row_t)`.