Added the touchtyping fixer feature.
This commit adds a new feature to fix your shift usage when touchtyping. It does this by checking whether a keycode uses the correct shift and if it doesn't it will either send the keycode unshifted or not send it, depending on whether `TOUCHTYPING_FIXER_SUPRESSIVE` is defined. You can modify the checked keycodes by defining `LEFT_SHIFT_BLOCKED` and `RIGHT_SHIFT_BLOCKED`.